Add assembly listing and executable generation capability 🎉
Implemented the missing parts of the code generation process. Code can now be compiled into assemblers, assembled into an executable format, and executed. Also changed the 'integer literal' data type representation from i64 to i32 in multiple files to handle shorter integer values. New dependencies added for better assembly listing and handling errors during decoding.

+++ src/ast.rs
@@ -77,7 +77,7 @@ impl<'a> AstNode for Expr<'a> {
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum ExprKind<'a> {
-    IntegerLiteral(i64),
+    IntegerLiteral(i32),
     StringLiteral(&'a str),
     FunctionCall(FunctionCallData<'a>),
     BinaryExpr(Box<Expr<'a>>, BinaryOp, Box<Expr<'a>>),
