+++ src/parser.rs
@@ -338,7 +338,7 @@ fn parse_grouped_expr(input: Span) -> IResult<Span, Expr> {
 
 fn parse_int_lit(input: Span) -> IResult<Span, Expr> {
     let (input, num_span) = digit1(input)?;
-    let num = num_span.fragment().parse::<i64>().unwrap();
+    let num = num_span.fragment().parse::<i32>().unwrap();
 
     let expr = Expr {
         kind: ExprKind::IntegerLiteral(num),
